Refactor MainActivity into focused extension files
